reorganize callbacks
